requirements.txt
fastapi
uvicorn
langchain
transformers
torch
langchain-community
langchain_openai
python-dotenv